Day 2 capstone, Account Service/currentUser

+++ java/tenmo-server/src/main/java/com/techelevator/tenmo/dao/UserDAO.java
@@ -2,6 +2,7 @@ package com.techelevator.tenmo.dao;
 
 import com.techelevator.tenmo.model.User;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 public interface UserDAO {
@@ -14,5 +15,5 @@ public interface UserDAO {
 
     boolean create(String username, String password);
 
-	double viewBalance(int id);
+	BigDecimal viewBalance(int id);
 }
